feat : Service 확장성 고려 인터페이스 설계
CouponService 인터페이스 생성
ServiceConfig로 구성될 ServiceImpl선택
